Refactor components in settings into separate files
