+++ src/screens/settingscreen/LimitedPlatesComponent.tsx
@@ -1,0 +1,17 @@
+import React, { useState } from "react";
+import { Switch } from "react-native";
+
+const LimitedPlatesComponent = () => {
+  const [isEnabled, setIsEnabled] = useState(false);
+  const toggleSwitch = () => setIsEnabled((previousState) => !previousState);
+  return (
+    <Switch
+      trackColor={{ false: "#767577", true: "#81b0ff" }}
+      ios_backgroundColor="#3e3e3e"
+      onValueChange={toggleSwitch}
+      value={isEnabled}
+    />
+  );
+};
+
+export default LimitedPlatesComponent;
